After delaying the addition of a track, don't blindly free the ItdbTrack
RbIpodDb is used to delay modifications of the ipod database while it's
marked as read-only. It's done with an Action datastructure which contains
all the information needed to perform the delayed action. This structure
then has to be freed. In the case of delayed addition of tracks to the
database, we were unconditionnally calling itdb_track_free on the content
of the Action, which should never happen. For now, I added a warning to catch
the cases when it's called with a non-NULL track, it can be removed when
we are more confident that the freeing isn't needed

svn path=/trunk/; revision=6001