add status icon plugin
This plugin uses GtkStatusIcon when built with gtk 2.16 or newer
(fixing bug 349280); can be disabled as required (fixing bug 317982);
makes some aspects of the status icon behaviour configurable (fixing
bugs 158168, 324114, 331019).

Notification popups are now a bit more configurable: they can be
disabled, only shown when the main window is hidden, or always shown.

The status icon is a configurable too:  it can be always hidden, only
shown while a notification popup is visible, always shown, or it can
own the main window, providing minimize/close-to-tray.  The new
implementation of close-to-tray uses the same way of hiding the window
as everything else, fixing bug 551002.

The dontreallyclose plugin is no longer needed, so it has been removed.