xfade: make RBXFadeStream a GstBin, rather than having it hold one
Override send_event for the stream bin so that events sent to the bin go
to the element that provides the bin's src pad.  This means we can seek
on the bin, rather than having to dig out the pad to send the event to.

Simplify the stream reuse handler in the audio CD plugin by taking
advantage of this.