encoder: add RBEncoderFactory singleton for plugin hooks
Since a new encoder instance is created for each file being transferred,
it's tricky for plugins to make use of any hooks we might add there.
RBEncoderFactory acts as a well known place for hooks to be located.
The only hook we have here so far is 'prepare-source', which does the
same thing as the signal of that name on the player backends.