metadata: figure out media types for missing plugins
We can reliably guess what media type the missing decoder or demuxer
processes based on the GStreamer media type strings, so we use that
information to update our knowledge of the media types in the file.

We don't need to use the 'unknown-type' signal any more, since the
missing plugin messages convey the same information more usefully.

With this, we only run the codec installer for audio-only files
(bug #537113).