player-gst: pretend playing time is 0 during track changes (bug #584804)
Since we don't know exactly when playbin2 changes from one track to the
next, we can't emit the playing-stream signal at the right point, and we
can't answer playing position queries properly either.  So, we emit the
playing-stream signal when we set the new URI and pretend the playback
position is 0 (in the new stream) until the next timer tick.