rhythmdb: allow errors to be replaced by song entries (bug #539007)
We do normal imports by creating events with no entry type
specified (for some reason); a logic bug in the metadata load processing
code meant that in this case, an import error would never be deleted and
replaced with a regular song entry.  Now torrent downloads and other
types of transfer into the music library will eventually be imported
correctly.