artdisplay: use fuzzy matching to filter search results (bug #349066)
Rework the processing of Amazon search results to use the fuzzy string
matcher I added earlier.  Rather than trying to do exact or substring
matches in various combinations, we now do fuzzy matches for the album
name and each of the artist names in the search results (with a special
case for 'various artists' to help compilations).

To avoid getting the wrong album by the right artist, we reject results
where one of the matches is below a threshold.  If the result doesn't
have a title or an artist, we use a default match strength slightly above
this threshold.  We then combine the artist and album match strengths
and take the highest result, as long as it's above another threshold.