fix interaction with pulseaudio flat volume (bug #584958)
First, we count the number of volume changes that have been made, and
keep track of the last change to be applied to the pipeline.  Then,
when starting the pipeline, we apply any volume changes that were made
while the pipeline was not running.

The extra trick here is that when the audio sink provides volume
control, we ignore the first volume setting, which restores the volume
setting from gconf.  The reason we do this is that pulseaudio remembers
our volume for itself and applies it correctly, without modifying the
reference output volume.

This change also fixes the volume change notification in the playbin2
backend, which was using the wrong signal.