audiocd: remove action group from UI manager when disposing (bug 597706)
The action group previously was not removed from the UI manager.  This
meant that the 'reload' action remained, with a pointer to the old
(disposed) source object as the user data.  Removing it from the UI
manager when disposing one source causes the action to be recreated when
creating a new source, so it'll have the right user data.