mtp: huge rewrite to do all device operations in a separate thread
This moves all MTP device operations to a separate thread, including
uploads and downloads, so the main UI thread never gets blocked.

Noteworthy things: the source object is now created based on a raw
device, so it has to open the device.  The source destroys itself if the
device cannot be opened.

There is now a GStreamer MTP sink element that writes data to a
temporary file, then uploads the file when it receives an EOS event.

Fixes bug #534981.