xfade: set the pipeline to PLAYING state so it selects a clock
Since we mess around with the states of various things in the pipeline
directly, we have to lock the state of stream bins and the output bin so
pipeline state changes don't affect them.  Combined with the previous
commit, this gets visualization working properly with the crossfading
player backend.  It also means there's a clock available for other
plugins, which might be useful for displaying spectrum data.