Add magic to support selected-source actions
Actions that operate on the current selected source are pretty common,
and so we had a pretty common pattern where a signal handler would go
and find the selected source from the shell, check it was the right
type, and then unref the source before returning.  This moves all of
that (except the type check) into a GClosure wrapper, simplifying the
signal handler code.