removable-media: fix compile error with GUdev 147+
Since device numbers are now always 64bit, we can't stuff them into
pointers and use them as hash keys with g_direct_hash as pointers are
only guaranteed to be at least 32 bits.  Instead, we need to allocate
and store a new copy of the device number and use
uint64_hash/uint64_equal.