rhythmdb: don't use rhythmdb_add_timeout_commit for metadata changes (bug #603929)
The problem with this is that we don't want these changes to be synced
with the file, but if we call rhythmdb_commit for some other reason,
such as hiding a file that is no longer present in the library
directory, before the timeout expires, the changes will be committed
with the sync_changes flag set to TRUE, so we'll end up writing the data
back to the file.

Other calls to rhythmdb_add_timeout_commit are safe, as they're made for
entries that can't be synced (such as import errors or ignored files) or
for changes that need to be synced.