mtp: use media player source base class
The main tricky thing here is that we track the amount of free space on
the device, rather than asking it when creating the properties dialog.
To do this, we add a new action type to the MTP thread that just calls a
specified function on the device handler thread.  After any track
transfer or delete action, we add a callback to update our idea of the
amount of free space on the device.