lyrics: use a single lyrics search site list
Rather than having the list of sites in the UI file, in the
configuration dialog code, and in the search code, now we have a single
list that we use to do everything.

Since the LyricWiki search doesn't work and shows no sign of ever
working again, it has been removed from the list (bug #603876).