media-player: implement syncing of music, podcasts, and playlists
The sync UI presents options for syncing all music, all podcasts,
or individual playlists and podcast feeds.  This is displayed on a
new 'sync' tab in the media player source properties dialog.

The sync process works by generating track identifiers (based on
several track metadata fields) for all tracks on the device, and
all tracks from the selected local playlists and podcasts.  It
compares the sets of track identifiers to determine which tracks
need to be transferred to the device, and which need to be deleted
from it.  After updating the files on the device, it then
reconstructs playlists on the device to match the local playlists
selected for syncing.

Sync settings are stored in a keyfile named after the device serial
number.  The settings consist of sync categories (such as playlists
and podcasts) and sync groups within categories (such as individual
playlists or podcast feeds).

This is based on Paul Bellamy's summer of code project.