playbin2: fix track changes (bug #601524, bug #602957)
Rather than emitting the playing stream change signal immediately, and
pretending the playback position is 0 until the next tick, we wait until
either we get a playbin2-stream-changed signal or a position query
returns a position inside the first second.  This means the playback
position should always be consistent, and the track change is reflected
in the UI closer to when it actually occurs.

This also means we don't emit the playing-stream signal on a streaming
thread, which fixes some threading problems with a11y enabled.