metadata: remove the need to load metadata before saving
There's no need to load the metadata first, the tags from the file will
be available in the tag writing pipeline anyway.  When saving metadata,
we need a typefind element in the pipeline, since we don't already know
the file type.

We now have a method to reset the metadata state.  This previously
happened at the start of the loading process, but we don't necessarily
do that between save operations any more.