rhythmdb: use sortnames in property models (bug #133444)
For each type of property model, we now have an ordered list of
properties to use to determine the sort key string.  Where there is a
sortname version of a property, it is preferred to the display version.

Each property model entry now includes an index into this list, which
records the property that provided the current sort string for that
entry.

If any of the entries with a given artist name (for example) has an
artist sortname set, then that is used for the entry in the property
model.  If there are multiple different sortnames for a single property,
the results are undefined.

When an entry changes, the property model needs to check if any of the
properties in its sort string order have changed.  When a higher
priority sort string appears, or the current highest priority sort
string changes or is removed, the property model entry is reordered
accordingly.