media-player: explicitly set up media player source action group
The ipod and mtp plugins load UI XML files referencing the sync action
from the media player source action group, so it must be registered
before that happens.  Since there won't be an instance of a media player
source at that point, we need a function to explicitly register the
action group.