media-player: rework sync settings a bit
Instead of the separate 'sync all music' and 'sync all podcasts'
settings, now everything is controlled from the tree view.
If the category item ('music' or 'podcasts') is enabled, everything in
the category will be synced.  Otherwise, only the enabled groups will
be.  If there are groups enabled, the category will be displayed as
inconsistent.

There's a new group in the 'music' category: 'all music', which
corresponds to the library source.  If enabled, all music will be
synced.  When this is enabled, the playlist items just control whether
to create playlists on the device.