replaygain: new plugin using GStreamer's ReplayGain elements
This plugin uses the rgvolume and rglimiter elements to process
ReplayGain tags and adjust the output volume accordingly.

It provides configuration options for selecting the gain mode ('radio'
or 'album'), setting the pre-amp level, and enabling or disabling
compression in order to prevent clipping.

It determines the fallback gain (used for tracks that don't provide
ReplayGain tags) by calculating the mean of the last 10 non-fallback
gain levels.