(Mostly) enable downloads for download accounts.
The purchase confirmation dialog has been removed, since we don't really need
any information from the user. There are 2 functions that seem to have been
removed from the rb module, which are required for the download to complete
successfully.

Also, use sku in name of partially downloaded file and in_progress file,
because the basename for current downloads is always api-download.php.