status-icon: use direction-neutral templates when necessary (bug #609767)
When the localised template strings ("by x", "from y") are in one
direction but the artist and album strings being substituted in are in
the other, the localised templates will look weird and wrong.  Instead,
we use direction neutral templates, just separating the album and artist
with a slash, so what's displayed at least makes sense.