audiocd: add checkbox column for selecting tracks to extract (bug #537483)
The 'extract' column controls the set of tracks that are extracted when
the 'extract to library' action is invoked.  The state of the checkbox is
stored in entry type specific data.

The column header includes a checkbox, and clicking on that toggles its
state, and sets the state of all tracks to match the new state.
Individual tracks can be toggled by clicking on the checkbox in the
entry view.