podcast: fix download cancellation
We don't need to report cancellation as a download error, since the user
is obviously aware of it.  When exiting the download thread, we should
always close the read and write streams whether the download was
cancelled or not, so we just don't pass in the cancellable to those
operations.