podcast: add all new posts to the database (bug #551402)
Now when updating a podcast feed, we add all posts that don't already
have entries, rather than only adding them if the post time is newer than
that of the most recent post.

When deleting a post, we just hide the entry, rather than deleting it, so
it doesn't get recreated next time the feed is updated.