[ipod] rework iPod database saving
When it has been modified, the saving of the iPod database is made
asynchronously after 15 seconds with no change. When lots of tracks are
copied to an iPod, this saving is triggered every 15 seconds during the
transfer which can slow the transfer down (especially on iPhone-like
devices). It makes more sense to only save the database once all transfers
are done, and not to wait 15 seconds for the final save, but 2 seconds.
With a 15 seconds delay, people don't wait long enough before unplugging
their device (especially on iOS devices)