[mtp] ignore mounts not associated with a device node
The mtp plugin tries to find the GMount associated with a given unix device
node, but it doesn't ignore GMount which do not have an associated device
node which can happen with an iPhone for example. Not ignoring these causes
the crash described in bug #613715. This also fixes a small reference leak
in the "mounts" list