podcast: rework post and feed deletion (bug #613784)
Instead of using the entry-deleted signal to delete downloaded posts and
remove posts for deleted feeds, we now do all that up front before
deleting anything.  This makes the code more straightforward, so
hopefully next time someone changes what happens when a podcast entry is
deleted, we won't forget to deal with half of it.