playbin2: do everything asynchronously
Rather than blocking until state changes finish, we now react to the
state change bus messages.  This of course greatly reduces the amount of
time we block the main loop, but also means that all signals are emitted
on the main thread, even in cases where we start playback from a
different thread (bug #614351).

Also it turns out we don't need to pause the pipeline before seeking, so
now we just send the seek event directly.