podcast: use webkit to display HTML episode descriptions (bug #320507)
The main interesting bit here is that we need to distinguish between
plain text and HTML descriptions, as interpreting plain text as HTML
removes all formatting.  Since we're dealing with small document
fragments here, normal content type detection doesn't work.  Instead, we
search for a set of strings, such as common tags and entities, that
reliably indicate HTML content.

The dependency on webkit is optional.