rhythmdb: replace entry-changed GSList of changes with a GValueArray
This is a bit inefficient - the change structures get copied twice while
preparing them for emission - but the new representation can be accessed
directly from python code.  This might slow startup by a tiny amount,
but otherwise won't be noticeable.

The various handlers we attach to entry-changed were all trivial to
convert.