podcast: use g_file_input_stream_query_info for downloads (bug #596615)
Where it works (depends on gvfs version), this means we only use a
single HTTP request for podcast downloads, rather than one to get the
filename and size and a second to download it.  Aside from being more
efficient, this also helps in cases where the server doesn't deal with
HEAD requests properly.