magnatune: Set names of radio buttons explicitly
In newer versions of pygtk(?), the name of a widget is not set from its id in
the builder file. So we need to iterate over the radio buttons in the prefs
dialog and set their name explicitly. We also need to use widget.get_name()
instead of widget.props.name to access the name.

Fixes bug #615658.