source: protect against repeated deletion (bug #613526)
Some devices appear and disappear when plugged in, which causes the
corresponding source to be deleted twice - once when the device
disappears and once when the device open operation fails.  The easy way
to fix this is just to check that a source hasn't already been deleted
when deleting it.