mtp: ignore devices that don't support any audio formats (bug #615743)
libmtp also detects PTP cameras. It does not actually know them (in the
udev rules), but since we do not check for this, PTP cameras regularly
appear in Rhythmbox, which is confusing and useless.

Check the supported file types of the device for audio types, and ignore
the device if there aren't any suitable ones. Please note that a lot of
cameras support read-only WAV, which is fairly uninteresting for our
purposes and would lead to detecting them as music players, so ignore WAV.