add new track transfer batch and queue system
This system groups track transfers into batches.  A batch consists of a
set of tracks to be transferred and the origin and destination sources.
Batches report overall progress, can be cancelled as a unit, and provide
a set of signals allowing greater integration into the UI, such as
indicating which track is currently being ripped from a CD.

The track transfer queue organises batches into an order of execution,
only processing one at a time.  It provides overall progress information
via signals and through a method.  It can also be used to find any
outstanding transfer batches with a given source as the origin or
destination.