audioscrobbler: authenticate using web services API
Preferences dialog has button to authorise user. When clicked, request
an "authorisation token". When the token is recieved send the user
to a url where they can authorise Rhythmbox, and add a timeout
function to the main loop which will continually check if we have
been authorised. This will then store a "session key" which is saved
and reused for future sessions.