audioscrobbler: login user-interface adjusts depending on status
A label informs the user if they are logged out, in the process of
logging in, or if they are successfully logged in. The button changes
between a log in, cancel, and log out button depending upon the status.
User interaction such as cancelling or logging out, or a response error
can cause the status to revert to logged out.