audioscrobbler: improve architecture of RBAudioscrobblerAccount
Seperate internal authentication code from user interface.
Provide public functions to authenticate and log out of an account.
Add "login-status-changed" signal which is emitted when an account logs
out, starts to log in, successfully logs in, or when there is an error
logging in. Adjust UI code to use new functions and signal.