fix various impl_get_status methods to not return progress 0.0
The annoying bit here is that Python implementations don't have a way to
decline to return a progress value, so for Magnatune and Jamendo we
return 2.0 (anything greater than 1.0 means 'hide the progress bar')
unless we have a specific value to return.  This means we lose the
progress value from rhythmdb for these sources, unfortunately.