add implementation of the MPRIS D-Bus spec
This is a new hidden plugin, enabled by default, using GDBus.
As a result, it is only built when glib 2.26 or newer is available.

The tracklist object has not been implemented.  It doesn't quite
fit into how rhythmbox works, so it probably won't be implemented
unless we need it for something interesting.