rhythmdb: process file move events (bug #620857)
When built against glib 2.24 or newer, we can request file move events
from the file monitor.  We handle file move events by just updating the
location of the entry, or deleting it if an entry already exists for the
new location.

While we're in there, simplify processing of file create/modify and delete
events.  We don't need to send the event through the rhythmdb event queue,
we can just make the required changes directly.  In the case of new files,
we can just pass the URI to rhythmdb_add_uri().

Based on a patch by Philipp Kerling <k.philipp@gmail.com>