rhythmdb: use correct entry types when retrying after plugin install
The entry types from the original metadata load event are still in the
event structure, so we should use them rather than defaulting to 'song'
type.  If the plugin install was triggered by scanning files on a
media player device, we don't want to import the file into the library.