rhythmdb: implement plugin installation in RhythmDBImportJob
At the end of an import job, we check if any of the import error entries
created included plugin installer detail strings, and if so, we emit
a missing-plugins signal for the plugin install code.  When the plugin
install finishes successfully, we retry the imports that resulted in
missing plugin errors.

This approach means that we don't process missing plugins for files imported
outside of a RhythmDBImportJob, which currently includes files picked up by
the library crawl on startup or by library monitoring.  This is probably
a good thing, on the whole.