rework missing-plugin helper to use direct function calls, not signals
Now it gets initialized with a GtkWindow on startup, but after that it can
be called directly from anywhere with just a list of plugin details and
a closure to call on completion.  This simplifies using it in import jobs,
and allows us to trigger plugin installation from the import error source.