import-errors: add an info bar allowing missing plugin install at any time
The info bar is only displayed when there is at least one import error in
the source that includes missing plugin information.  Clicking the button
runs the installer for all missing plugins required for files currently in
the import error source, and on successful install, attempts to import
all of the files again.