audioscrobbler: add initial radio streaming support
RBAudioscrobblerRadioSource is a source which populates an entry view
with entries fetched from a specific Last.fm radio station. These are
added to the source list as children of the profile source. In the future
the user will be able to create and delete these at will, but for the time
being only a few are added for testing purposes.

The source is currently capable of fetching and parsing the playlist that
Last.fm provides, and playing these tracks when they are selected.