mtp: create albums in the same folder as the tracks (bug #623667)
For many (most? all?) MTP devices to actually find the album data, which
includes the album art image, it needs to be created in the same folder
as the tracks themselves.  Since we create the album after uploading the
first track, we use folder ID from that track for the album too.