encoder: fix setup error reporting (bug #625124)
Rather than uselessly returning a boolean if the pipeline creation fails or
it refuses to change state, we asynchronously emit the 'completed' signal
(using an idle handler).  For state change failures, we wait until we get
an error message on the bus and handle that.  This simplifies error handling
in the transfer batch too.