rhythmdb: add an entry type method for processing availability changes
Availability changes are things like filesystems being mounted and unmounted,
results of checking for files on startup, and 'resource not found' errors when
trying to play an entry.  The exact behaviour we want varies between entry types,
and so far we've mostly just done what's right for normal song entries.