shell-player: convert 'not found' errors to entry availability events
If we get a 'not found' error from the player backend, we should update
the entry's availability just as if the file was missing during the startup
library check.  To avoid getting into a fight with the play order, we need to
do this after skipping to the next track.