rhythmdb: rework handling of files on local and remote mounts during startup
Rather than attempting to mount each time we find an entry on a missing mount,
we group entries by mount point while the database is loading.  Entries with
a NULL mountpoint are always checked on startup, as they're on core filesystems
that we can assume are fast and that will always be mounted.

Entries on local mountpoints are checked if the filesystem is mounted and
otherwise hidden.  We can't actually try to mount the filesystem if it's not
mounted, as we can't identify which volume they're on, and attempting to mount
all available volumes would be kind of obnoxious.  If the filesystem is mounted
later, all entries on it will be checked.

Entries on remote mountpoints (with URI schemes other than file://) are never
checked.  Instead, they're hidden if not mounted, shown otherwise.  We currently
don't try to automatically mount remote filesystems, but this will happen eventually.