Fix typeahead search to use the current sorted column (bug #329984)
For example, if you click the artist column (to sort by artist), typeahead
search will search by artist, rather than title which is the original static
default. If the user does not click a column, it will default to searching by
title. If the user clicks a non-string column like Time, it will also default
to using title.