shell: add a dbus method for activating a source by URI
Given a URI, this method searches for a source for which rb_source_uri_is_source
returns TRUE, then selects it, and optionally starts playback from it.  This is
intended to be used for newly inserted devices, but if we defined a URI scheme for
playlists and the like, it could also be used for those.