mediaserver2: initial implementation of MediaServer2 dbus spec
This is about the least interesting MediaServer2 implementation we could
possibly do, but here it is.  So far, it only exports the library, doesn't
provide any kind of album/artist/genre container structure, and doesn't
implement searching.

For now, the plugin is visible and disabled by default.  It will probably be
hidden and enabled by default when it's more complete.