encoder: add hardcoded extensions for common media types (bug #625054)
Previously we relied on the audio encoding profiles to provide extension
information, but of course that only works if the encoder elements for the
profile are available.  You shouldn't need to install an MP3 encoder to copy
MP3 files to a device, so now we have a fallback for the most common types.