[ipod] reworkd delayed playlist renaming
When playlist renaming was delayed, the previous code was trying to
store both the playlist to change and its name to the same place in
memory (because of the use of an union). This was obviously bad and
causing silent memory corruption.
To avoid that, use RbIpodDelayedPlaylistOp to store information
about the playlist renaming since this structure is big enough to
hold both the playlist and its new name.