[ipod] fix addition/removal of playlist items (bug #601152)
The code that was connecting the addition/removal callbacks to the
playlist model wasn't connecting to the correct model, so the
callbacks were never getting called, and thus rhythmbox was never
actually telling libgpod to add new data items to the playlists.
This is easily fixed by using the correct model, though we'd also need
to handle notify::base-query-model since it can change.