shell-player: cancel playlist parsing when stopping (bug #612156)
If the user tries to stop playback of a radio stream before the playlist
parser has finished, we need to cancel the parser and ignore any URLs we
get from it so we don't try to start playback again.  Similarly, if the
user tries to start playback while the parser is still going, we can just
ignore that.