podcast: add child sources showing new episodes and recent downloads
This splits the podcast source in two, into a base class that provides the feed
and post views and implements the various UI actions, and a subclass that
contains the parts unique to the main source, mostly the configuration dialog.

The child sources are populated by database queries.  The new episode source
uses the post-time property and the new download source uses the modification
time, which previously wasn't set at all for podcast posts, but now is set to
the current time when the download starts.

The queries are currently hardcoded to match episodes posted and downloaded in
the last week, but we could provide hour/day/week/month selections if that
seems useful enough.

The child sources currently use the same gconf keys as the main source, so they
have the same sort order and browser settings.