status-icon: make use of some new notification hints and capabilities
"action-icons" is the official spec name for "x-gnome-action-buttons".

We also use the "resident" hint to request that the server keep our
notification around after the bubble (or banner) has expired, so we can
just update it rather than creating a new one.

When action buttons are supported, we switch between "play" and "pause"
depending on whether we're currently playing.  When an appropriate hint
is added to the notification spec, we'll update this when we start or
stop playback, but for now it'd be a bit too annoying.