podcast: fix podcast status on download completion (bug #633469)
Only set the status to 'downloaded' if we've downloaded the whole file.

If a selected entry is in the download queue, enable the 'cancel download'
action.  If a selected entry is not in the download queue and has a status
other than 'downloaded', enable the 'download episode' action.

This should make it possible to complete downloads that fail due to
connection errors and the like.