replace entry equality checks with entry location equality checks (bug #634957)
It appears that at some point we switched from doing a pointer
comparison when no specific compare method existed to just returning
False.  As a result, we need to compare entry locations instead.
This broke lyrics display in the context pane plugin and cover art in
the mpris plugin, among other things.