entry-view: relax sorting column assertion
Sorting keys are now bound to settings during RBBrowserSource
construction, and subclasses haven't had a chance to insert any
custom columns at that point.  So, we need to allow the entry view
to go unsorted for a while.  The sort order will be applied when
the relevant column is inserted.