xfade: fix race condition with preroll vs end of previous stream (bug #646405)
Previously there was a window where a stream would not be started
if it finished prerolling after the previous stream reached EOS, but
before the EOS message was processed on the main thread, and the
new stream needed to wait for the previous stream to finish.

We now check for new streams to start both when the stream reaches EOS
(processed on the stream thread) and when the EOS bus message is
processed (on the main thread), so the new stream will always be started.
If it doesn't preroll in time, there may be a slight gap in the audio
output, but that's unavoidable if preroll is too slow.