track-transfer-batch: use lossless if that's all we have (bug #661806)
If the encoding target for the batch only includes lossless profiles,
use the first one that works.  The only situation this should occur in
is when extracting from an audio CD with FLAC as the preferred media type.

Otherwise lossless profiles are still ignored, even if a lossy profile
requires additional plugins to be installed.